Cleaned up parser package
