+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/metadata/SpecimenCollection.java
@@ -1,4 +1,6 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.metadata;
+
+import io.github.programminglife2016.pl1_2016.parser.JsonSerializable;
 
 import java.util.Collection;
 import java.util.Iterator;
@@ -6,15 +8,15 @@ import java.util.Iterator;
 /**
  * A data structure that represents the segments.
  */
-public interface NodeCollection extends JsonSerializable, Iterable<Node> {
+public interface SpecimenCollection extends JsonSerializable, Iterable<Subject> {
     /**
      * Add a segment to the collection.
      *
      * @param id      id of the segment
-     * @param node the actual segment
+     * @param specimen the actual segment
      * @return the segment parameter
      */
-    Node put(Integer id, Node node);
+    Subject put(String id, Subject specimen);
 
     /**
      * Get a segment by id.
@@ -22,7 +24,7 @@ public interface NodeCollection extends JsonSerializable, Iterable<Node> {
      * @param id id of the segment
      * @return segment
      */
-    Node get(Object id);
+    Subject get(Object id);
 
     /**
      * Checks if the collection contains a particular key.
@@ -36,13 +38,13 @@ public interface NodeCollection extends JsonSerializable, Iterable<Node> {
      * Return all segments.
      * @return all segments.
      */
-    Collection<Node> getNodes();
+    Collection<Subject> getSpecimen();
 
     /**
      * Iterator for the collection.
      * @return iterator
      */
-    default Iterator<Node> iterator() {
-        return getNodes().iterator();
+    default Iterator<Subject> iterator() {
+        return getSpecimen().iterator();
     }
 }
