+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/phylotree/PhyloGeneticTreeParser.java
@@ -1,4 +1,7 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.phylotree;
+
+import io.github.programminglife2016.pl1_2016.parser.JsonSerializable;
+import io.github.programminglife2016.pl1_2016.parser.Parser;
 
 import java.io.InputStream;
 import java.util.ArrayList;
