+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/phylotree/TreeNode.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.phylotree;
 
 import java.util.List;
