+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/segments/NodeMap.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.segments;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -6,18 +6,13 @@ import java.util.HashMap;
 /**
  * Adapter for segment hashmap
  */
-public class SpecimenMap extends HashMap<String, Subject> implements SpecimenCollection {
+public class NodeMap extends HashMap<Integer, Node> implements NodeCollection {
 
-    /**
-     * Create segment hashmap.
-     */
-    public SpecimenMap() {
-    }
     /**
      * Create segment hashmap.
      * @param initialCapacity capacity of the hashmap.
      */
-    public SpecimenMap(int initialCapacity) {
+    public NodeMap(int initialCapacity) {
         super(initialCapacity);
     }
 
@@ -26,7 +21,7 @@ public class SpecimenMap extends HashMap<String, Subject> implements SpecimenCol
      *
      * @return all segments
      */
-    public Collection<Subject> getSpecimen() {
+    public Collection<Node> getNodes() {
         return values();
     }
 }
