+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/segments/NodeSerializer.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.segments;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
