+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/segments/PositionManager.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.segments;
 
 /**
  * Manager for calculating postions of nodes. Classes implementing this interface
