+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/SegmentCollectionTest.java
@@ -4,6 +4,8 @@ package io.github.programminglife2016.pl1_2016.parser;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
+import io.github.programminglife2016.pl1_2016.parser.segments.NodeCollection;
+import io.github.programminglife2016.pl1_2016.parser.segments.Segment;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
