+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/SegmentListTest.java
@@ -2,6 +2,7 @@
 
 package io.github.programminglife2016.pl1_2016.parser;
 
+import io.github.programminglife2016.pl1_2016.parser.segments.NodeList;
 import org.junit.Before;
 
 /**
