+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/SegmentParserTest.java
@@ -1,5 +1,8 @@
 package io.github.programminglife2016.pl1_2016.parser;
 
+import io.github.programminglife2016.pl1_2016.parser.segments.Node;
+import io.github.programminglife2016.pl1_2016.parser.segments.NodeCollection;
+import io.github.programminglife2016.pl1_2016.parser.segments.SegmentParser;
 import org.junit.Before;
 import org.junit.Test;
